refactor test folders
