+++ __tests__/__examples__/MixedTests.example.ts
