+++ __tests__/__integration__/render/suites/TestSuite.test.ts
@@ -2,10 +2,10 @@ import { TestSuite } from "../../../../src/render/suites/TestSuite";
 import { Test } from "../../../../src/render/tests/Test";
 import { IResultsProcessorInput } from "../../../../src/processor/doc/jest/IResultsProcessorInput";
 
-const simplePassingTests: IResultsProcessorInput = require("../../../data/simplePassingTests.json");
-const simpleFailingTests: IResultsProcessorInput = require("../../../data/simpleFailingTests.json");
-const twoSideBySideFailDiffTests: IResultsProcessorInput = require("../../../data/twoSideBySideFailDiffTests.json");
-const largeTests: IResultsProcessorInput = require("../../../data/largeTests.json");
+const simplePassingTests: IResultsProcessorInput = require("../../../__resources__/simplePassingTests.json");
+const simpleFailingTests: IResultsProcessorInput = require("../../../__resources__/simpleFailingTests.json");
+const twoSideBySideFailDiffTests: IResultsProcessorInput = require("../../../__resources__/twoSideBySideFailDiffTests.json");
+const largeTests: IResultsProcessorInput = require("../../../__resources__/largeTests.json");
 
 describe("TestSuite tests", () => {
