+++ __tests__/__resources__/addedSnapshotTests.json
