+++ __tests__/__resources__/removedSnapshotTests.json
