+++ __tests__/utils/Logger.test.ts
@@ -1,4 +1,4 @@
-import { Logger } from "../../../src/utils/Logger";
+import { Logger } from "../../src/utils/Logger";
 
 describe("Logger tests", () => {
     it("should get a log instance", () => {
