+++ web/template.html
@@ -52,10 +52,10 @@
                         </div>
                     </li>
                 </ul>
-                <form class="form-inline mt-2 mt-md-0">
+                <!-- <form class="form-inline mt-2 mt-md-0">
                     <input class="form-control mr-sm-2" type="text" placeholder="Search" aria-label="Search" disabled>
                     <button class="btn btn-outline-success my-2 my-sm-0" type="submit" disabled>Search</button>
-                </form>
+                </form> -->
             </div>
         </nav>
     </header>
